src/main/java/BankAccount.java
//the class
    public class BankAccount {

    private String firstName;
    private String lastName;
    private String dateOfBirth;
    private String accountNumber;   //this could be a string because int is used when doing calculations
    private int balance;
    private int overDraft;



    //Down below is the Constructor to intialisze the BankAccount Object
    public BankAccount(String firstName, String lastName, String dateOfBirth,
                       String accountNumber, int balance) {
        this.firstName = firstName;
        this.lastName = lastName;
        this.dateOfBirth = dateOfBirth;
        this.accountNumber = accountNumber;
        this.balance = balance;
    }

    public int getOverDraft() {
        return overDraft;
    }

    public void setOverDraft(int overDraft) {
        this.overDraft = overDraft;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }


    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getDateOfBirth() {
        return dateOfBirth;
    }

    public void setDateOfBirth(String dateOfBirth) {
        this.dateOfBirth = dateOfBirth;
    }

    public String getAccountNumber() {
        return accountNumber;
    }

    public void setAccountNumber(String accountNumber) {
        this.accountNumber = accountNumber;
    }

    public int getBalance() {
        return balance;
    }

    public void setBalance(int balance) {
        this.balance = balance;
    }

    public void deposit(int amount){
         this.balance += amount;
    }

    public void withdrawal(int amount){
        this.balance -= amount;
    }
}




